refactor: token refresh and logout moved to auth
